Always increment server playerId

+++ src/server/EA/PlasmaSession.cs
@@ -11,6 +11,6 @@ public class PlasmaSession
     public required string ClientString { get; init; }
     public required string PartitionId { get; init; }
 
-    public int PID { get; set; }
+    public long PID { get; set; }
     public long EGAM_TID { get; set; }
 }
